carrot 2.4.28
git-svn-id: http://svn.b-shock.co.jp/commons/forms/trunk@4259 6bc21610-e001-0410-9273-fe8b54c00d2d

+++ lib/carrot/database/record/BSRecord.class.php
@@ -8,7 +8,7 @@
  * テーブルのレコード
  *
  * @author 小石達也 <tkoishi@b-shock.co.jp>
- * @version $Id: BSRecord.class.php 2343 2010-09-13 03:38:22Z pooza $
+ * @version $Id: BSRecord.class.php 2394 2010-10-16 08:57:37Z pooza $
  * @abstract
  */
 abstract class BSRecord implements ArrayAccess,
@@ -170,11 +170,13 @@ abstract class BSRecord implements ArrayAccess,
 			throw new BSDatabaseException($this . 'を削除することはできません。');
 		}
 
-		foreach ($this->getTable()->getChildClasses() as $class) {
-			$table = BSTableHandler::getInstance($class);
-			$table->getCriteria($this->getTable()->getName() . '_id', $this);
-			foreach ($table as $record) {
-				$record->delete();
+		if (!$this->getDatabase()->hasForeignKey()) {
+			foreach ($this->getTable()->getChildClasses() as $class) {
+				$table = BSTableHandler::getInstance($class);
+				$table->getCriteria($this->getTable()->getName() . '_id', $this);
+				foreach ($table as $record) {
+					$record->delete();
+				}
 			}
 		}
 		if ($record = $this->getParent()) {
@@ -243,7 +245,7 @@ abstract class BSRecord implements ArrayAccess,
 	/**
 	 * 親レコードを返す
 	 *
-	 * 適切にオーバーライドするば、update等の動作が少し利口に。
+	 * 適切にオーバーライドすれば、update等の動作が少し利口に。
 	 *
 	 * @access public
 	 * @return BSRecord 親レコード
